+++ packages/components/src/Filter/FilterBar/_docs/FilterBar.stories.tsx
@@ -21,7 +21,7 @@ import { FilterBar, Filters, useFilterBarContext } from "../index"
 import { FilterBarMultiSelectProps } from "../subcomponents"
 
 const meta = {
-  title: "Components/Filter Bar",
+  title: "Components/FilterBar",
   component: FilterBar,
   argTypes: {
     filters: { control: false },