+++ packages/components/src/Filter/FilterBar/context/types.ts
@@ -28,6 +28,7 @@ export type FilterBarStateFilters<ValuesMap extends FiltersValues> = {
 
 export type FilterBarState<ValuesMap extends FiltersValues> = {
   hasUpdatedValues: boolean
+  hasRemovableFilter: boolean
   filters: FilterBarStateFilters<ValuesMap>
   activeFilterIds: Set<keyof ValuesMap>
   values: Partial<ValuesMap>