+++ packages/components/src/Filter/FilterDateRangePicker/subcomponents/FilterDateRangePickerField/FilterDateRangePickerField.tsx
@@ -259,6 +259,10 @@ export const FilterDateRangePickerField = ({
       return
     }
 
+    if (state.inputStartValue === "" && state.inputEndValue === "") {
+      return
+    }
+
     const newStartDate = validateStartDate(
       selectedRange?.from,
       state.inputStartValue