+++ packages/components/src/Tile/subcomponents/GenericTile/GenericTile.tsx
@@ -1,4 +1,5 @@
-import React, { HTMLAttributes, useState } from "react"
+import React, { HTMLAttributes, useState, useRef, useEffect } from "react"
+import { useIntl } from "@cultureamp/i18n-react-intl"
 import classnames from "classnames"
 import { AllowedHeadingTags, Heading } from "~components/Heading"
 import { Text } from "~components/Text"
@@ -22,6 +23,8 @@ export type GenericTileProps = {
   titleTag?: AllowedHeadingTags
   metadata?: string
   information?: TileInformation | React.ReactNode
+  /** Provides accessible label for the title's info button @default "View more information: [title]" */
+  infoButtonLabel?: string
   /** @deprecated Use `variant` instead */
   mood?:
     | "positive"
@@ -46,6 +49,7 @@ export const GenericTile = ({
   titleTag = "h3",
   metadata,
   information,
+  infoButtonLabel,
   mood,
   variant = "default",
   footer,
@@ -53,6 +57,35 @@ export const GenericTile = ({
   ...restProps
 }: GenericTileProps): JSX.Element => {
   const [isFlipped, setIsFlipped] = useState<boolean>(false)
+  const [isDocumentReady, setIsDocumentReady] = useState<boolean>(false)
+
+  const { formatMessage } = useIntl()
+  const infoButtonRef = useRef<HTMLButtonElement>(null)
+  const infoButtonReturnRef = useRef<HTMLButtonElement>(null)
+
+  useEffect(() => {
+    setIsDocumentReady(true)
+  }, [])
+
+  useEffect(() => {
+    if (!isDocumentReady) {
+      setIsDocumentReady(true)
+      return
+    }
+
+    if (isFlipped) {
+      infoButtonReturnRef.current!.focus()
+    } else {
+      infoButtonRef.current!.focus()
+    }
+  }, [isFlipped])
+
+  const translatedInfoLabel = formatMessage({
+    id: "kzGenericTile.infoButtonLabel",
+    defaultMessage: "View more information:",
+    description:
+      "Prompts user to interact with button to reveal more information",
+  })
 
   const renderTitle = (): JSX.Element => (
     <div className={styles.title}>
@@ -80,11 +113,12 @@ export const GenericTile = ({
       {information && (
         <div className={styles.informationBtn}>
           <IconButton
-            label="Information"
+            label={infoButtonLabel || `${translatedInfoLabel} ${title}`}
             icon={<Icon name="info" isPresentational isFilled />}
             onClick={(): void => setIsFlipped(true)}
             disabled={isFlipped}
             aria-hidden={isFlipped}
+            ref={infoButtonRef}
           />
         </div>
       )}
@@ -135,15 +169,22 @@ export const GenericTile = ({
   const renderBack = (): JSX.Element | void => {
     if (!information) return
 
+    const returnButtonLabel = formatMessage({
+      id: "kzGenericTile.infoButtonReturnLabel",
+      defaultMessage: "Hide information:",
+      description: "Prompts user to interact with button to hide information",
+    })
+
     return (
       <div className={classnames(styles.face, styles.faceBack)}>
         <div className={styles.informationBtn}>
           <IconButton
-            label="Information"
+            label={`${returnButtonLabel} ${title}`}
             icon={<Icon name="arrow_back" isPresentational />}
             onClick={(): void => setIsFlipped(false)}
             disabled={!isFlipped}
             aria-hidden={!isFlipped}
+            ref={infoButtonReturnRef}
           />
         </div>
         <div className={styles.information}>